# marketstack-openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Marketstack OpenAPI",
    "description": "Inofficial marketstack OpenAPI spec",
    "version": "1.0"
  },
  "paths": {
...
    "/intraday": {
      "get": {
        "tags": [
          "intraday"
        ],
        "summary": "Query",
        "operationId": "intraday",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Intraday interval",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Interval"
                }
              ]
            },
            "example": "5min",
            "name": "interval",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.IntervalPrice__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dividends": {
      "get": {
        "tags": [
          "dividends"
        ],
        "summary": "Query",
        "operationId": "dividends",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.Dividend__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tickers": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Query",
        "operationId": "tickers",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search string",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.Ticker__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers/{symbol}": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol",
        "operationId": "ticker_symbol",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticker"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers/{symbol}/eod": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Eod",
        "operationId": "ticker_symbol_eod",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_TickerEod_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers/{symbol}/eod/latest": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Eod Latest",
        "operationId": "ticker_symbol_eod_latest",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EodPrice"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers/{symbol}/eod/{date}": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Eod Date",
        "operationId": "ticker_symbol_eod_date",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": true,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EodPrice"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers/{symbol}/intraday": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Intraday",
        "operationId": "ticker_symbol_intraday",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Intraday interval",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Interval"
                }
              ]
            },
            "example": "5min",
            "name": "interval",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_TickerIntraday_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers/{symbol}/intraday/latest": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Intraday Latest",
        "operationId": "ticker_symbol_intraday_latest",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntervalPrice"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "title": "ErrorResponse",
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Interval": {
        "title": "Interval",
        "enum": [
          "1min",
          "5min",
          "10min",
          "30min",
          "1hour",
          "3hour",
          "6hour",
          "12hour",
          "24hour"
        ],
        "description": "An enumeration."
      },
...
      "PagedResponse_List_models.Dividend__": {
        "title": "PagedResponse[List[models.Dividend]]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "title": "Data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dividend"
            }
          }
        }
      },
...
      "PagedResponse_List_models.IntervalPrice__": {
        "title": "PagedResponse[List[models.IntervalPrice]]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "title": "Data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntervalPrice"
            }
          }
        }
      },
...
      "PagedResponse_List_models.Ticker__": {
        "title": "PagedResponse[List[models.Ticker]]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "title": "Data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ticker"
            }
          }
        }
      },
...
      "Sort": {
        "title": "Sort",
        "enum": [
          "ASC",
          "DESC"
        ],
        "description": "An enumeration."
      },
...
    }
  }
}